+++ R/global_vars.R
@@ -1,7 +1,7 @@
 # defining global variables and functions to appease R CMD Check
 
 utils::globalVariables(
-  names = c(".", "rowid", "isanoutlier"),
+  names = c(".", "rowid", "df"),
   package = "ipmisc",
   add = FALSE
 )